% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{followup}
\alias{followup}
\title{Follow-up}
\format{
A data frame with 93 rows and 19 variables:
\describe{
  \item{\code{ID}}{character. Participant ID}
  \item{\code{thoughtquitting}}{character. In the last week, how often have you thought about quitting your tobacco use?}
  \item{\code{tobaccotoday}}{character. How many times have you used tobacco today? }
  \item{\code{quit}}{character. Have you successfully quit using any type of tobacco? If so, please explain.}
  \item{\code{consequence}}{character. In the last week, how often have you thought about the consequences of using tobacco?}
  \item{\code{craved}}{character. In the last week, how often have you craved tobacco?}
  \item{\code{encourage}}{character. In the last week, has anyone encouraged you to use tobacco? If yes, who?}
  \item{\code{discourage}}{character. In the last week, has anyone discouraged you from using tobacco? If yes, who?}
  \item{\code{cuttingback}}{character. In the last week, how often have you thought about cutting back on your tobacco use?}
  \item{\code{methods}}{character. What methods have you used to cut back and/or quit tobacco use?}
  \item{\code{presentation_share}}{character. Have you told anyone about what you learned in the presentation about tobacco use?}
  \item{\code{presmemory1}}{character. Please tell me what content you remember from the presentation.}
  \item{\code{presmemory2}}{character. Please tell me what content you remember from the presentation.}
  \item{\code{presmemory3}}{character. Please tell me what content you remember from the presentation.}
  \item{\code{presmemory4}}{character. Please tell me what content you remember from the presentation.}
  \item{\code{presmemory5}}{character. Please tell me what content you remember from the presentation.}
  \item{\code{NOTES}}{character. Additional notes}
  \item{\code{Date}}{double. Date of baseline interview and sample collection}
  \item{\code{Date2}}{double. Date of followup interview and sample collection}
  \item{\code{delta_days}}{double. Number of days between baseline and followup}
  \item{\code{quit2}}{character. Did you quit using tobacco?}
  \item{\code{thoughtquitting2}}{double. Have you thought of quitting tobacco?}
  \item{\code{tobaccotoday2}}{double. Did you use tobacco today}
  \item{\code{cuttingback2}}{character. Have you thought of cutting back on tobacco use?}
  \item{\code{presentation_share2}}{character. Did you share information from the presentation with anyone?}
  \item{\code{harm1}}{character. Type of harm (general, reproductive, or emic)}
  \item{\code{harm2}}{character. Type of harm (general, reproductive, or emic)}
  \item{\code{harm3}}{character. Type of harm (general, reproductive, or emic)}
  \item{\code{harm4}}{character. Type of harm (general, reproductive, or emic)}
  \item{\code{harm5}}{character. Type of harm (general, reproductive, or emic)}
  \item{\code{number_general}}{integer. DESCRIPTION.}
  \item{\code{number_reproductive}}{integer. DESCRIPTION.}
  \item{\code{number_emic}}{integer. DESCRIPTION.}
  \item{\code{specific_harm1}}{character. First free-listed harm at followup}
  \item{\code{specific_harm2}}{character. Second free-listed harm at followup}
  \item{\code{specific_harm3}}{character. Third free-listed harm at followup}
  \item{\code{specific_harm4}}{character. Fourth free-listed harm at followup}
  \item{\code{specific_harm5}}{character. Fifth free-listed harm at followup}
  }
}
\usage{
followup
}
\description{
Interviews after collection of second sample
}
\keyword{datasets}
