% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples}
\alias{samples}
\title{Sample cotinine concentrations}
\format{
A data frame with 142 rows and 13 variables:
\describe{
  \item{\code{Plate}}{double. Plate number.}
  \item{\code{Assay}}{double. Assay number.}
  \item{\code{ID}}{character. Subject ID.}
  \item{\code{pre_post}}{integer. Pre: pre-intervention cotinine concentration. Post: post-intervention concentration.}
  \item{\code{OD1}}{double. Optical density of well 1.}
  \item{\code{OD2}}{double. Optical density of well 2.}
  \item{\code{Mean}}{double. Mean OD}
  \item{\code{NSB}}{double. Non-specific binding. The mean OD of the two wells with only buffer.}
  \item{\code{Zero}}{double. The mean OD of the two wells with cotinine concentration = 0.}
  \item{\code{B_B0}}{double. The fraction of bound cotinine.}
  \item{\code{group_id}}{integer. Assay group number.}
  \item{\code{cotinine}}{double. Estimate cotinine concentration (ng/ml).}
}
}
\usage{
samples
}
\description{
Salivary cotinine concentrations (ng/ml) assayed with
Salimetrics ELISA cotinine kits.
}
\keyword{datasets}
